log optimization and error handling

+++ intro/pics/picDifference.py
@@ -29,46 +29,53 @@ def delete_file(fileName):
 jpgFilePath = input('Enter JPG File path: ')
 # print('JPG File Path: {}' .format(jpgFilePath))
 jpgFileNameList = glob.glob(jpgFilePath + '/*.JP*')
-print_list(jpgFileNameList)
+# print_list(jpgFileNameList)
+print('[{}] JPG files present.' .format(len(jpgFileNameList)))
 
 # get NEF files from file system
 nefFilePath = input('Enter NEF File path: ')
 # print('NEF File Path: {}' .format(nefFilePath))
 nefFileNameList = glob.glob(nefFilePath + '/*.NEF')
-print_list(nefFileNameList)
+# print_list(nefFileNameList)
+print('[{}] NEF files present.' .format(len(nefFileNameList)))
 
 
 deletedFiles = []
-# Delete extra NEF Files
-n = 0
-for nefFullFileName in nefFileNameList:
-    n = n+1
-    print('-----')
-    # basaename returns only filename without full Path
-    # splitext returns filename till last . (i.e. remove extension)
-    nefFileName = os.path.splitext(os.path.basename(nefFullFileName))[0]
-    print('NEF File-{}: Searching file [{}] in JPG List' .format(n, os.path.basename(nefFileName)))
 
-    j = 0
-    found = False
-    # search for nefFile in jpgFileNameList
-    for jpgFullFileName in jpgFileNameList:
-        j = j+1
-        jpgFile = os.path.splitext(os.path.basename(jpgFullFileName))[0]
-        print('JPG File-{}: Basename [{}] ' .format(j, jpgFile))
+if len(jpgFileNameList) > 0:
+    # Delete extra NEF Files only if JPG files list is Valid, and not Blank
+    n = 0
+    for nefFullFileName in nefFileNameList:
+        n = n+1
+        print('-----')
+        # basaename returns only filename without full Path
+        # splitext returns filename till last . (i.e. remove extension)
+        nefFileName = os.path.splitext(os.path.basename(nefFullFileName))[0]
+        print(
+            'NEF File-{}: Searching file [{}] in JPG List' .format(n, os.path.basename(nefFileName)))
 
-        if(nefFileName == jpgFile):
-            # delete NEF file from file system
-            print('Match Found, DO NOT DELETE: Moving to next NEF file !')
-            found = True
-            break
-        else:
-            print('NO Match: Searching next JPG File')
+        j = 0
+        found = False
+        # search for nefFile in jpgFileNameList
+        for jpgFullFileName in jpgFileNameList:
+            j = j+1
+            jpgFile = os.path.splitext(os.path.basename(jpgFullFileName))[0]
+            # print('JPG File-{}: Basename [{}] ' .format(j, jpgFile))
 
-    if not found:
-        print('NO Match: Deleting NEF File [{}]' .format(nefFullFileName))
-        deletedFiles.append(nefFullFileName)
-        delete_file(nefFullFileName)
+            if(nefFileName == jpgFile):
+                # delete NEF file from file system
+                print('Match Found, DO NOT DELETE: Moving to next NEF file !')
+                found = True
+                break
+            # else:
+            #     print('NO Match: Searching next JPG File')
+
+        if not found:
+            print('NO Match: Deleting NEF File [{}]' .format(nefFullFileName))
+            deletedFiles.append(nefFullFileName)
+            delete_file(nefFullFileName)
 
 print('Total [{}] Files Deleted!!' .format(len(deletedFiles)))
-print(deletedFiles)
+
+if len(deletedFiles) > 0:
+    print(deletedFiles)
